+++ src/main.ts
@@ -270,7 +270,7 @@ document.addEventListener("keydown", (event): void => {
     } else if (event.key === "Backspace") {
         // Delete selected nodes
         for (let i = GRAPH.nodes.length - 1; i >= 0; i--) {
-            if (GRAPH.nodes[i].selected) {
+            if (GRAPH.nodes[i].selected && GRAPH.nodes[i].div.getAttribute("contenteditable") === "false") {
                 GRAPH.nodes[i].delete();
             }
         }
